refactor: `BottomSheet`으로 컴포넌트명 변경 및 리팩터링
